src/requirements.txt
pydantic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
requests
PyMuPDF